+++ main/src/me/ricky/kbot/core/data/MongoDBHandler.kt
@@ -1,0 +1,5 @@
+package me.ricky.kbot.core.data
+
+class MongoDBHandler {
+
+}