+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 0.0.0
+#### 0.0.0
 * Command Handlers, MultiCommandHandler, ServerCommandHandler
 * Prefix Handler
 * Some Utilities
+
+#### 0.0.1
+* Added BotConfig Serializable
+* Added KBot class to make bot instances
+* some other tweaks